requirements.txt
tensorflow==2.12.0
opencv-python==4.7.0.72
numpy==1.23.5
pillow==9.5.0
matplotlib==3.7.1
scikit-learn==1.2.2
pytesseract==0.3.10
flask==2.3.2
ultralytics==8.0.120
supervision==0.6.0
